block full application launch while running tests
see: https://pointfreeco.github.io/swift-dependencies/main/documentation/dependencies/testing#Testing-gotchas

+++ Mlem/MlemApp.swift
@@ -8,6 +8,7 @@
 import SwiftUI
 import UIKit
 import Nuke
+import XCTestDynamicOverlay
 
 @main
 struct MlemApp: App {
@@ -19,40 +20,42 @@ struct MlemApp: App {
 
     var body: some Scene {
         WindowGroup {
-            Window(selectedAccount: accountsTracker.defaultAccount)
-                .environmentObject(accountsTracker)
-                .onAppear {
-                    var imageConfig = ImagePipeline.Configuration.withDataCache(name: "main", sizeLimit: AppConstants.cacheSize)
-                    imageConfig.dataLoadingQueue = OperationQueue(maxConcurrentCount: 8)
-                    imageConfig.imageDecodingQueue = OperationQueue(maxConcurrentCount: 8) // Let's use those CORES
-                    imageConfig.imageDecompressingQueue = OperationQueue(maxConcurrentCount: 8)
-                    
-                    // I'm leaving that here for mormaer, once I get a handle on rate limites that's where we put em!
-                    // imageConfig.isRateLimiterEnabled
-                    ImagePipeline.shared = ImagePipeline(configuration: imageConfig)
-                    
-                    URLCache.shared = AppConstants.urlCache
-
-                    setupAppShortcuts()
-                    
-                    // clear out tmp directory
-                    let tempDirectory = FileManager.default.enumerator(atPath: FileManager.default.temporaryDirectory.absoluteString
-                    )
-                    if let tempDirectory = tempDirectory {
-                        for case let file as URL in tempDirectory {
-                            try? FileManager.default.removeItem(at: file)
+            if !_XCTIsTesting {
+                Window(selectedAccount: accountsTracker.defaultAccount)
+                    .environmentObject(accountsTracker)
+                    .onAppear {
+                        var imageConfig = ImagePipeline.Configuration.withDataCache(name: "main", sizeLimit: AppConstants.cacheSize)
+                        imageConfig.dataLoadingQueue = OperationQueue(maxConcurrentCount: 8)
+                        imageConfig.imageDecodingQueue = OperationQueue(maxConcurrentCount: 8) // Let's use those CORES
+                        imageConfig.imageDecompressingQueue = OperationQueue(maxConcurrentCount: 8)
+                        
+                        // I'm leaving that here for mormaer, once I get a handle on rate limites that's where we put em!
+                        // imageConfig.isRateLimiterEnabled
+                        ImagePipeline.shared = ImagePipeline(configuration: imageConfig)
+                        
+                        URLCache.shared = AppConstants.urlCache
+                        
+                        setupAppShortcuts()
+                        
+                        // clear out tmp directory
+                        let tempDirectory = FileManager.default.enumerator(atPath: FileManager.default.temporaryDirectory.absoluteString
+                        )
+                        if let tempDirectory = tempDirectory {
+                            for case let file as URL in tempDirectory {
+                                try? FileManager.default.removeItem(at: file)
+                            }
                         }
+                        
+                        // set app theme to user preference
+                        let windowScene =  UIApplication.shared.connectedScenes.first as? UIWindowScene
+                        windowScene?.windows.first?.overrideUserInterfaceStyle = lightOrDarkMode
+                        
+                        let appearance = UITabBarAppearance()
+                        appearance.backgroundEffect = UIBlurEffect.init(style: .systemThinMaterial)
+                        UITabBar.appearance().standardAppearance = appearance
+                        UITabBar.appearance().scrollEdgeAppearance = appearance
                     }
-
-                    // set app theme to user preference
-                    let windowScene =  UIApplication.shared.connectedScenes.first as? UIWindowScene
-                    windowScene?.windows.first?.overrideUserInterfaceStyle = lightOrDarkMode
-
-                    let appearance = UITabBarAppearance()
-                    appearance.backgroundEffect = UIBlurEffect.init(style: .systemThinMaterial)
-                    UITabBar.appearance().standardAppearance = appearance
-                    UITabBar.appearance().scrollEdgeAppearance = appearance
-                }
+            }
         }
         .onChange(of: lightOrDarkMode) { value in
             let windowScene =  UIApplication.shared.connectedScenes.first as? UIWindowScene
